requirements.txt
streamlit==1.32.2
PyYAML==6.0.1
openai==1.68.2


httpx
pyvis